src/requirements.txt
numpy
pandas
openai
nltk
stemming
spacy
re
greek-stemmer-pos
cltk